# openapi.yaml
openapi: 3.0.3
info:
  title: JIRA Clone API
  description: |
    A comprehensive API for managing projects, issues, and users in a JIRA-like application.
    
    ## Features
    - Project management with team collaboration
    - Issue tracking with Kanban-style workflow
    - Real-time updates and notifications
    
  version: 1.0.0
  contact:
    name: JIRA Clone Team
    email: support@jira-clone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development serverxs
paths:
...
  /comments/issues/{issueId}/comments:
    get:
      operationId: getIssueComments
      tags:
        - Comments
      summary: Get comments for an issue
      description: Returns all comments for a specific issue with pagination support
      parameters:
        - name: issueId
          in: path
          description: Issue ID
          required: true
          schema:
            type: string
            format: objectId
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of comments per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createComment
      tags:
        - Comments
      summary: Create a new comment
      description: Creates a new comment for the specified issue
      parameters:
        - name: issueId
          in: path
          description: Issue ID
          required: true
          schema:
            type: string
            format: objectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  description: Comment content
                  example: "This is a comment on the issue"
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comment:
                        $ref: '#/components/schemas/Comment'
                  message:
                    type: string
                    example: "Comment created successfully"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{id}:
    get:
      operationId: getComment
      tags:
        - Comments
      summary: Get a specific comment
      description: Returns details of a specific comment
      parameters:
        - name: id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
            format: objectId
      responses:
        '200':
          description: Comment retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comment:
                        $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateComment
      tags:
        - Comments
      summary: Update a comment
      description: Updates the content of a comment (author only)
      parameters:
        - name: id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
            format: objectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  description: Updated comment content
                  example: "This comment has been updated"
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comment:
                        $ref: '#/components/schemas/Comment'
                  message:
                    type: string
                    example: "Comment updated successfully"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Can only edit your own comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteComment
      tags:
        - Comments
      summary: Delete a comment
      description: Deletes a comment (author or admin only)
      parameters:
        - name: id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
            format: objectId
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Comment deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Can only delete your own comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/issues/{issueId}/comments/stats:
    get:
      operationId: getCommentStats
      tags:
        - Comments
      summary: Get comment statistics for an issue
      description: Returns statistics about comments for a specific issue
      parameters:
        - name: issueId
          in: path
          description: Issue ID
          required: true
          schema:
            type: string
            format: objectId
      responses:
        '200':
          description: Comment statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      stats:
                        $ref: '#/components/schemas/CommentStats'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Issue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/author/{authorId}:
    get:
      operationId: getCommentsByAuthor
      tags:
        - Comments
      summary: Get comments by author
      description: Returns all comments created by a specific author with pagination
      parameters:
        - name: authorId
          in: path
          description: Author user ID
          required: true
          schema:
            type: string
            format: objectId
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of comments per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          description: Unique user identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        role:
          type: string
          enum: [admin, project_manager, developer, designer, qa, viewer]
          default: viewer
          description: User's role in the system
          example: "developer"
        isActive:
          type: boolean
          default: true
          description: Whether the user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:30:00.000Z"

    # Project Schemas
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          description: Unique comment identifier
          example: "507f1f77bcf86cd799439011"
        issueId:
          type: string
          format: objectId
          description: ID of the issue this comment belongs to
          example: "507f1f77bcf86cd799439010"
        author:
          oneOf:
            - type: string
              format: objectId
            - $ref: '#/components/schemas/User'
          description: User who wrote this comment
          example: "507f1f77bcf86cd799439012"
        content:
          type: string
          description: Comment content
          example: "This issue has been resolved in the latest deployment."
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: Attachments associated with this comment
        createdAt:
          type: string
          format: date-time
          description: Comment creation timestamp
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Comment last update timestamp
          example: "2024-01-15T11:30:00.000Z"

    # Comment Statistics Schema
    CommentStats:
      type: object
      properties:
        totalComments:
          type: integer
          description: Total number of comments for the issue
          example: 15
        totalAttachments:
          type: integer
          description: Total number of attachments across all comments
          example: 3
        uniqueAuthors:
          type: integer
          description: Number of unique authors who have commented
          example: 5

    # Error Schema
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "User already exists"
...